Refresh the lstopo mode and add a generic "command" mode. Change the lstopo command to add a dash '-' which tells it to give text based output rather than graphical, also extend this mode to allow the precise lstopo command to be specified as an option.  The command given is exected on the target node and %p is replaced with the pid of the target process.  This should future proof padb as when a --pid option gets added to lstopo padb will be able to use it by just editing a configuration file rather than editing the code.
Also add a generic "command" mode which will run any command, again
substituting %p with the pid of the target process, this is in effect
what lstopo is now so add a mode to do this specifically othewise I'm
sure people would try and piggy-back this behaviour onto the lstopo
mode.  The default command is 'readlink /proc/%p/exe'.