Extend padb to work with (a subset of) LSF jobs.
Note that this commit does not allow padb to attach to all LSF jobs but rather
padb is now able to detect two types of wrapper script commonly used with LSF and 
extract the information it needs from the wrapper script.  As such this is a step
forward but not a general case solution for LSF support and will not meet
everybody's need.

This commit is largely the work of Thipadin @ Bull as sent to the developers
list on 9/2/10 with the following changes:

The resource manager has been renamed to the more simple 'lsf'.
lsf_mpi has been replaced with lsf_mode with valid values being mpich2
 or openmpi
slurp_remote_cmd() has been added instead of calling slurm_cmd() with
 "ssh host ..."