Fail better in the light of new kernel security features preventing the use of ptrace on non-child processes.
It seems new kernels (Ubuntu 10.10 included) have a file /proc/sys/kernel/yama/ptrace_scope
which controls which processes can be ptraced, with this param at it's default
value programs can only ptrace their children which stops gdb from
attach to programs.  If we fail to attach ever then check if this file exists
and if it's value is non-zero then tell the user it's a likely cause of the problem.