Add a "launch_mode" to control how the backend processes are launched.
The four options are local,ssh,pdsh and rmgr and are set as a comma seperated
list which is walked until one that is able to launch the job is found.

local: launch the inner process locally.  Only works for single node jobs on the local host.
ssh: launch the inner process using ssh.  Only works for single node jobs.
pdsh: launch using pdsh.
rmgr: launch using the selected resource manager.

The old behaviour was equilivent to a setting of "rmgr,local,pdsh", the new
default it "local,rmgr,ssh,pdsh".