Add support for using clush as a backend launcher. This is similar to the pdsh support but uses clush instead. http://sourceforge.net/projects/clustershell/
The mechanism for launch is the same as pdsh so presumably the performance
is the same at scale, it might suit some people to use clush rather than
pdsh though.  Pdsh will still be used in preference but clush will be
selected if pdsh cannot be found or if the user selects it.