Fix the deadlock detection code.
There were some changes to the handling of minfo output and the collective
state had not been updated so wasn't being reported correctly, and hence
the deadlock detection code had nothing to work with.  Alter the minfo
output to be in the new format and correctly match/forward this in the padb
inner processes.