Update the gdb/mi parsing code.
Re-instate gdb_strip_value() to a older version to avoid problems elsewhere
in the code but rather modify the value of "value" in the gdb_var_addr()
function as this is where the output of gdb has changed.

Handle a value of "<optimized out>" for a pointer and do not attempt parse
it as a real pointer.